Completed First EDA.
This was indeed a fun way to practice python, pandas and numpy. More work can be done this Analysis, Majority of it has been completed. Minor things such as Grammar, Decorations(Visually Appealing), Few Missing Analytics such as noting the accuracy, testing across all values of minor conditions, etc. It is ready to be shared on my socials, a few tweaks(hence the last release, after this commit) should be posted.